requirements.txt
bs4
requests
openai
paho-mqtt